+++ src/index.js
@@ -1,6 +1,7 @@
 import { AuthenticaionService } from "./services/auth.service";
 import { ExtractionService } from "./services/extraction.service";
 import { DropboxService } from "./services/dropbox.service";
+import { ReviewTimerService } from "./services/review-timer.service";
 
 const shared = {
   counter: undefined,
@@ -25,6 +26,7 @@ const shared = {
 
 const extractionService = new ExtractionService(shared);
 const dropboxService = new DropboxService(shared);
+const reviewTimerService = new ReviewTimerService(shared);
 
 //! Dropbox credentials
 
@@ -42,43 +44,7 @@ let dbx  = auth.createDbxConnection();
 // ----------------------------------------------------------------------------------------
 
 //Don't check token on review page
-let startTimer;
-if ( window.location.pathname.includes("generate_review") || 
-window.location.pathname.includes("generate_dfe_review")) {
-  createUI();
-
-  // Start Review Timer
-  loadTimer();
-  startTimer = setInterval(() => reviewTimer(), 1000);
-}
-
-//Reset timer
-timeResetIcon.addEventListener("click", async () => {
-  console.log(`%c timer reset`, 'color: #ffba08')
-  clearInterval(startTimer);
-
-  shared.counter = 0;
-  shared.min = 0;
-  localStorage.setItem("minutes", null);
-  localStorage.setItem("counter", null);
-  shared.combinedTime = 0;
-
-
-  startTimer = setInterval(() => reviewTimer(), 1000);
-  await loadTimer();
-
-});
-
-//===================================================== review count
-
-// Increment review count when review is finished and save to local storage
-reviewCompleteBtn?.addEventListener("click", () => {
-  shared.reviewCount++;
-  localStorage.setItem("reviewCount", shared.reviewCount);
-
-  counterEl.style.color = "#8BC34A";
-  counterEl.style.animationDuration = "3s";
-});
+reviewTimerService.setUpTimer(createUI)
 
 // ----------------------------------------------------------------------------------------
 // ----------------------------SET UP BUTTON STUFF ----------------------------------------
@@ -349,144 +315,6 @@ function highlightTaskName(taskName) {
 }
 
 //====================================================Review Timer
-async function loadTimer() {
-  /*
-
-   When the program is closed, it saves the current counter and
-    time to local storage using the beforeunload event, 
-    and clears the interval to stop incrementing the counter. 
-    When the program is restarted, it initializes the counter 
-    and last saved time from local storage again, and repeats 
-    the process to include the missing seconds and continue 
-    incrementing the counter.
-   */
-
-  if (window.location.pathname.includes("generate_review")) {
-    // Start the interval when the page is visible
-    async function handleVisibilityChange() {
-      if (document.visibilityState === "visible") {
-        clearInterval(startTimer);
-        startTimer = setInterval(() => reviewTimer(), 1000);
-        await loadTimer();
-        document.removeEventListener(
-          "visibilitychange",
-          handleVisibilityChange
-        ); //remove event listener to avoid multiple executions
-      } else {
-        clearInterval(startTimer);
-      }
-    }
-
-    document.addEventListener("visibilitychange", handleVisibilityChange);
-  } else {
-    clearInterval(startTimer);
-  }
-
-
-  // Initialize the counter and last saved time from local storage, or use default values
-  shared.counter = parseInt(localStorage.getItem("counter")) || 0;
-  shared.min = parseInt(localStorage.getItem("minutes")) || 0;
-  shared.lastSavedTime =
-    parseInt(localStorage.getItem("lastSavedTime")) || new Date().getTime();
-
-  // Calculate the elapsed time since the last saved time and add it to the counter
-  shared.counter += Math.floor((new Date().getTime() - shared.lastSavedTime) / 1000);
-
-  // Save the current counter and time to local storage when the program is closed
-
-  window.addEventListener("beforeunload", async () => {
-    saveTimeValues();
-    clearInterval(startTimer);
-  });
-
-
-
-  //From the "Review Submit" page, reset timer when returning to dashboard
-  let myWord = "Return to dashboard";
-  let aTags = document.querySelectorAll("a");
-  aTags?.forEach((item) => {
-    if (item?.textContent.includes(myWord)) {
-      item.addEventListener("click", () => {
-        shared.counter = 0;
-        shared.min = 0;
-        localStorage.setItem("minutes", null);
-        localStorage.setItem("counter", null);
-        localStorage.setItem("lastSavedTime", null);
-        clearInterval(startTimer);
-        if (localStorage.getItem("rememberReview") == "false") {
-          localStorage.setItem("id_improve_comments", "");
-          localStorage.setItem("id_positive_comments", "");
-          localStorage.setItem("id_overall_comments", "");
-        }
-      });
-    }
-  });
-}
-
-function reviewTimer() {
-  shared.counter++;
-
-  // Check if the value of 'counter' has exceeded 59 seconds
-  if (shared.counter > 59) {
-    // Calculate the number of full minutes that have elapsed
-    // and add it to the 'min' variable
-    // The Math.floor() method is used to round down to the nearest integer
-    // e.g. if 'counter' is 75 seconds, the quotient of counter / 60 is 1.25,
-    // but we only want to add 1 full minute to 'min'
-    shared.min += Math.floor(shared.counter / 60);
-
-    // Calculate the remaining seconds after removing the full minutes
-    // and set the value of 'counter' to this value
-    // The modulus operator (%) returns the remainder of the division
-    // e.g. if 'counter' is 75 seconds, the remainder of counter % 60 is 15,
-    // which represents the number of seconds that have elapsed after 1 full minute
-    shared.counter %= 60;
-  }
-
-  //set time warning colors
-  if (shared.min < 5) {
-    counterEl.style.color = "#8BC34A";
-    counterEl.style.animationDuration = "10s";
-  } else if (shared.min < 7) {
-    counterEl.style.color = "#ffeb3b";
-    counterEl.style.animationDuration = "5s";
-  } else if (shared.min < 11) {
-    counterEl.style.color = "#ff9800";
-    counterEl.style.animationDuration = "2s";
-  } else if (shared.min < 13) {
-    counterEl.style.color = "#ff5722";
-    counterEl.style.animationDuration = "1s";
-  } else if (shared.min < 15) {
-    counterEl.style.color = "#f44336";
-    counterEl.style.animationDuration = ".2s";
-  } else {
-    counterEl.style.color = "#f44336";
-    counterEl.style.animationDuration = ".2s";
-  }
-
-  //Combine minutes and seconds into one string
-  shared.combinedTime = `${shared.min > 9 ? "" : 0}${shared.min}:${shared.counter > 9 ? "" : 0}${shared.counter}`;
-
-  //Stop timer at 60 minutes
-  if (shared.min > 60) {
-    counterEl.style.color = "#f44336";
-    counterEl.style.animationDuration = ".2s ";
-    shared.combinedTime = "59:00";
-    clearInterval(startTimer);
-  }
-
-  counterEl.textContent = shared.combinedTime;
-
-  // Save the current counter and time to local storage every second
-  saveTimeValues();
-}
-
-// Save the current counter and time to local storage
-function saveTimeValues() {
-  localStorage.setItem("counter", shared.counter);
-  localStorage.setItem("minutes", shared.min);
-  localStorage.setItem("lastSavedTime", new Date().getTime());
-}
 
 // Get review count from local storage and display it
 function getReviewCounts() {
